Moving audit script and phishing controller to app/scripts
