+++ package.json
@@ -8,7 +8,7 @@
     "dev:brave": "yarn build dev",
     "dist:brave": "yarn build prod:brave",
     "publish:brave": "yarn dist:brave && yarn build brave-publish",
-    "audit-deps": "node brave/scripts/audit.js",
+    "audit-deps": "node app/scripts/audit.js",
     "build": "node development/build/index.js",
     "start:test": "yarn build testDev",
     "benchmark:chrome": "SELENIUM_BROWSER=chrome node test/e2e/benchmark.js",
